Include some initial DB design and schema files
We are working with dbdiagram.io as the design tool of choice for this
project, so included in the repository is now the current DBML file
that describes the database as we currently envision it.

In addition, we include also the start of the SQL files that will define
and populate the actual database.

One of these sets up the schema, and is intended to be executed against
any database and ensure that the proper schema is in place, dropping any
data and tables required to do so (unconditionally).

The second bootstraps the initial data into the tables in order to get
things up and running. Currently this is the known set of roles and also
an initial user that represents me.