Update schema to set the default role of users
In order to make it easier to insert new users without having to fetch
the list of possible roles, set the default role when it's not
specified to be the user role.

In addition, for clarity ensure that items that are boolean not only
have a BOOLEAN type and a name that signals that they are boolean, but
also a boolean default value.

D1 supports boolean values going into the database, where the SQLite
underlay converts them to integers. If we add a layer on the output
that turns such fields back into boolean, everything should be cleaner
to understand.