Include some common DB fetching logic
This is a spin on the helpers created for Luddatumbazo, with
enhancements that build the query and the call to the logger that also
splits the results out into a single call.

In addition, rather than having to manually prepare all statements, the
helpers now take an array of SQL strings, array values to bind, and
previously compiled statements, building the list of statements
directly.

This new system is shorter, and also makes it easier to build batches of
statements and to reuse the same statement as a part of a batch by just
binding different values to it each time.

Lastly, since it is such a common paradigm, there is now a synthesized
fetchOne type option which gathers the results and then returns only
the first item, or null if none was found.

This is synthesized rather than using the core D1 API because D1 doesn't
support fetching only a single result without also dropping the metadata
for some reason.

This could be synthesized as needed by adding LIMIT 1 to any query to
ensure that the result set stops after the first result returned.