Remove CORS and implicit auth support from backend
We are playing with the notion of having the worker contain only the
server API and the ability to use Acheron to get information about the
current user without supporting login or logout in any fashion.

This will be pushed into the client facing code instead, which will bind
to us in order to carry things out.

FOr this reason, our route bindings are also removed, since the API will
be interacted with only via pages functions.