Add auth code and worker proxy to the client side
This completes what was started in the previous commit and shifts all of
the auth code to happen directly at the client side, and also includes a
proxy to shunt all API calls to the dosiero backend api directly.

This requires some extra values to be configured when deploying the UI
since it now also needs to know about auth.