Include some simple DB logic to handle user info
The authentication middleware has been updated so that it will try to
find the internal userId of the user that is currently authenticated
for the request, and if no such id can be found it will add a new user
to the database.

This allows the user database to dynamically track the users who have
logged in without requiring an explicit step.

This still requires some caching to make things a bit faster. Also it
does not support the notion of blocked users.