Fix issue with new users with no email address
The database enforces that the email address should always be specified
even if it's just an empty string (which is the default).

This constraint gets violated when the Acheron backend is github and the
user logging in has not specified an email address to be public in
their profile.

In such a case, we need to ensure that an email address gets provided.