Display avatar in place of the generic user icon
Small change, this updates the UI side to display the profile image for
the currently logged in user, if they have one. When there is not one,
a generic image is displayed instead.

This needs to be further refined because this currently will not work
for people that don't have a profile image defined (unless GitHub
ensures that everyone has one, anyway).

Also, probably when you're not logged in something more distinct than a
generic profile should be displayed. This is just proof of concept
though.