Set up a better layout for the package.json files
Largely this sets up the root of the project to have a package.json file
so that it can bootstrap the installation of all of the subfolders, so
spinning up a new development environment from the top is easier.

In addition to this, since the client side owns Acheron now, scripts
were set up in the client package.son for all acheron scripts, and the
one at the top level defers to that one to execute commands.